Split api into api and dnccontract.
